+++ NitroxServer/Communication/Packets/Processors/DiscordRequestIPProcessor.cs
@@ -1,4 +1,5 @@
-﻿using System.Net;
+﻿using System;
+using System.Net;
 using System.Threading.Tasks;
 using NitroxModel.Helper;
 using NitroxModel.Packets;
@@ -32,15 +33,34 @@ public class DiscordRequestIPProcessor : AuthenticatedPacketProcessor<DiscordReq
 
     private async Task ProcessPacketAsync(DiscordRequestIP packet, Player player)
     {
-        Task<IPAddress> wanIp = NetHelper.GetWanIpAsync();
-
-        if (await wanIp == null)
+        string result = await GetIpPortAsync();
+        if (result == "")
         {
             Log.Error("Couldn't get external Ip for discord request.");
             return;
         }
 
-        packet.IpPort = ipPort = $"{wanIp.Result}:{serverConfig.ServerPort}";
+        packet.IpPort = ipPort = $"{result}:{serverConfig.ServerPort}";
         player.SendPacket(packet);
+    }
+
+    /// <summary>
+    /// Get the WAN IP address or the Hamachi IP address if the WAN IP address is not available.
+    /// </summary>
+    /// <returns>Found IP or blank string if none found</returns>
+    private static async Task<string> GetIpPortAsync()
+    {
+        Task<IPAddress> wanIp = NetHelper.GetWanIpAsync();
+        Task<IPAddress> hamachiIp = Task.Run(NetHelper.GetHamachiIp);
+        if (await wanIp != null)
+        {
+            return wanIp.Result.ToString();
+        }
+
+        if (await hamachiIp != null)
+        {
+            return hamachiIp.Result.ToString();
+        }
+        return "";
     }
 }
